DTB: The old DTB burns my screen. Must replace it.
The first recovery with the old DTB can boot in fastboot, but soon, my
screen turned to be BURNT. When I replaced the DTB with a new one coming
with TWRP 3.0.3, everything is OK.

Neither the kernel nor the ramdisk has bug. A DTB with wrong config can
directly do harm to actual hardwares, just like the evil virus CIH decades
ago.